Moved the html_to_markdown and generate_markdown_from_text methods to markdown_helper.py under utils, Modified docx_to_markdown.py and url_to_markdown.py accordingly
